% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrr.R
\name{constrr}
\alias{constrr}
\title{Constructor Function for Hypothesis Test}
\usage{
constrr(x, y, alpha = double(), paired = logical())
}
\arguments{
\item{x}{vector sample from First population}

\item{y}{vector sample from second population}

\item{alpha}{a value that is within the range of 0 and 1}

\item{paired}{a logical value of either True or False}
}
\value{
List of data containing several outputs
}
\description{
This functions main purpose is to
}
\details{
The constructor makes an object and tests for NULL Hypothesis to
verify if the underlying means of two populaions are equal
}
\examples{
Constrr(x,y)
}
