% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrr.R
\name{myttest}
\alias{myttest}
\title{Constructor Function for Hypothesis Test

This functions main purpose is to

The constructor makes an object and tests for NULL Hypothesis to
verify if the underlying means of two populations are equal}
\usage{
myttest(x, y, alpha = double(), paired = logical())
}
\arguments{
\item{x}{vector sample from First population}

\item{y}{vector sample from second population}

\item{alpha}{a value that is within the range of 0 and 1}

\item{paired}{a logical value of either True or False}
}
\value{
List of data containing several outputs
}
\description{
Constructor Function for Hypothesis Test

This functions main purpose is to

The constructor makes an object and tests for NULL Hypothesis to
verify if the underlying means of two populations are equal
}
\examples{

x=rnorm(20,mean=10,sd=15);y=rnorm(10,mean=8,sd=15);constrr(x,y,0.05,FALSE)
}
