% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotmethod.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Plot method to create boxplot}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{Object from the constructor function "myttest"}

\item{...}{additional values for plotting}
}
\value{
Box Pot
}
\description{
Plot method to create boxplot
}
