% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Printmethod.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Method Function for printing the COnfidence Interval}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x...Object}{as input}
}
\value{
list showing confidence interval
}
\description{
Method Function for printing the COnfidence Interval
}
